refactor(web): migrate bridge feature to standard architecture
Refactors the bridge feature to comply with the feature architecture standard established in #6936.

Changes:
- Add public API barrel (index.ts) with lazy-loaded Bridge component
- Extract constants (BRIDGE_WIDGET_URL, LOCAL_STORAGE_CONSENT_KEY) to constants.ts
- Create barrel files for hooks and components
- Remove redundant inner dynamic() import in Bridge component
- Update external imports to use public API

Benefits:
- Proper code splitting at feature boundary
- Consistent structure with other features
- Simplified lazy loading (single boundary instead of nested)
- Improved maintainability

Resolves: https://linear.app/safe-global/issue/WA-1259/refactor-bridge-feature-to-new-feature-architecture
