+++ apps/web/src/features/bridge/components/Bridge/index.tsx
@@ -1,19 +1,10 @@
-import dynamic from 'next/dynamic'
-
 import { AppRoutes } from '@/config/routes'
 import { FeatureWrapper } from '@/components/wrappers/FeatureWrapper'
 import { SanctionWrapper } from '@/components/wrappers/SanctionWrapper'
 import { DisclaimerWrapper } from '@/components/wrappers/DisclaimerWrapper'
 import { FEATURES } from '@safe-global/utils/utils/chains'
-
-const LOCAL_STORAGE_CONSENT_KEY = 'bridgeConsent'
-
-const BridgeWidget = dynamic(
-  () => import('@/features/bridge/components/BridgeWidget').then((module) => module.BridgeWidget),
-  {
-    ssr: false,
-  },
-)
+import { LOCAL_STORAGE_CONSENT_KEY } from '../../constants'
+import { BridgeWidget } from '../BridgeWidget'
 
 export function Bridge() {
   return (
