+++ apps/web/src/features/bridge/components/index.ts
@@ -1,0 +1,2 @@
+export { Bridge } from './Bridge'
+export { BridgeWidget } from './BridgeWidget'
