+++ apps/web/src/features/bridge/constants.ts
@@ -1,0 +1,2 @@
+export const BRIDGE_WIDGET_URL = 'https://iframe.jumper.exchange/bridge'
+export const LOCAL_STORAGE_CONSENT_KEY = 'bridgeConsent'
