+++ apps/web/src/features/bridge/hooks/index.ts
@@ -1,0 +1,2 @@
+export { useIsBridgeFeatureEnabled } from './useIsBridgeFeatureEnabled'
+export { useIsGeoblockedFeatureEnabled } from './useIsGeoblockedFeatureEnabled'
