+++ apps/web/src/pages/bridge.tsx
@@ -1,7 +1,7 @@
 import Head from 'next/head'
 import type { NextPage } from 'next'
 
-import { Bridge } from '@/features/bridge/components/Bridge'
+import Bridge from '@/features/bridge'
 import { BRAND_NAME } from '@/config/constants'
 
 const BridgePage: NextPage = () => {
