+++ specs/002-bridge-refactor/checklists/requirements.md
@@ -1,0 +1,37 @@
+# Specification Quality Checklist: Bridge Feature Refactor
+
+**Purpose**: Validate specification completeness and quality before proceeding to planning  
+**Created**: 2026-01-15  
+**Feature**: [spec.md](../spec.md)
+
+## Content Quality
+
+- [x] No implementation details (languages, frameworks, APIs)
+- [x] Focused on user value and business needs
+- [x] Written for non-technical stakeholders
+- [x] All mandatory sections completed
+
+## Requirement Completeness
+
+- [x] No [NEEDS CLARIFICATION] markers remain
+- [x] Requirements are testable and unambiguous
+- [x] Success criteria are measurable
+- [x] Success criteria are technology-agnostic (no implementation details)
+- [x] All acceptance scenarios are defined
+- [x] Edge cases are identified
+- [x] Scope is clearly bounded
+- [x] Dependencies and assumptions identified
+
+## Feature Readiness
+
+- [x] All functional requirements have clear acceptance criteria
+- [x] User scenarios cover primary flows
+- [x] Feature meets measurable outcomes defined in Success Criteria
+- [x] No implementation details leak into specification
+
+## Notes
+
+- The spec references specific file paths (`index.ts`, `types.ts`, etc.) which are necessary for a refactoring task but don't constitute implementation details—they describe the required structure, not how to implement it
+- Success criteria SC-004 mentions "chunk file" which is a build output, not an implementation detail—it's a measurable outcome
+- The spec correctly preserves the existing geoblocking integration as a requirement rather than changing it
+- All requirements are derived from the established standard in 001-feature-architecture, ensuring consistency across feature migrations
