+++ specs/002-bridge-refactor/contracts/README.md
@@ -1,0 +1,11 @@
+# Contracts: Bridge Feature Refactor
+
+This directory is empty because the bridge feature refactor does not introduce any API contracts.
+
+The bridge feature is a pure UI refactoring that:
+
+- Reorganizes file structure to match the standard pattern
+- Creates barrel files for public API
+- Updates import paths
+
+No new APIs, endpoints, or programmatic interfaces are created.
